requirements.txt
blinker==1.6.2
click==8.1.7
Flask==3.0.0
imglyb==2.1.0
itsdangerous==2.1.2
jgo==1.0.5
Jinja2==3.1.2
JPype1==1.4.1
labeling==0.1.13
MarkupSafe==2.1.3
numpy==1.26.0
packaging==23.2
pandas==2.1.1
Pillow==9.5.0
psutil==5.9.5
pyimagej==1.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.3
scyjava==1.9.1
six==1.16.0
tifffile==2023.9.26
tzdata==2023.3
Werkzeug==3.0.0
xarray==2023.9.0
